Improve SVG warnings
