+++ src/utils/image.ts
@@ -59,7 +59,7 @@ async function loadImage (url: string): Promise<HTMLImageElement> {
  * @param {string} blob Image
  * @returns {Promise<HTMLImageElement>} Promise, which resolves into HTMLImageElement
  */
-async function blobToImg(blob: Blob): Promise<HTMLImageElement> {
+export async function blobToImg(blob: Blob): Promise<HTMLImageElement> {
     const url = URL.createObjectURL(blob);
     try {
         return await loadImage(url);
@@ -68,54 +68,27 @@ async function blobToImg(blob: Blob): Promise<HTMLImageElement> {
     }
 }
 
-async function drawableFromBlob (blob: Blob): Promise<HTMLImageElement|ImageBitmap> {
-    // Special-case SVG. We need to avoid createImageBitmap because of
-    // https://bugs.chromium.org/p/chromium/issues/detail?id=606319.
-    if (blob.type.startsWith('image/svg+xml')) {
-        // Firefox throws if you try to draw an SVG to canvas that doesn't have width/height.
-        // In Chrome it loads, but drawImage behaves weirdly.
-        // This function sets width/height if it isn't already set.
-        const parser = new DOMParser();
-        const text = await new Response(blob).text();
-        const doc = parser.parseFromString(text, 'image/svg+xml');
-        const svg = doc.documentElement;
-
-        if (doc.getElementsByTagName('parsererror').length > 0) throw Error('This is not a valid SVG');
-
-        if (svg.hasAttribute('width') && svg.hasAttribute('height')) {
-            return blobToImg(blob);
-        }
-
-        const viewBox = svg.getAttribute('viewBox');
-        if (viewBox === null) throw Error('SVG must have width/height or viewBox');
-
-        const [, , w, h] = viewBox.split(/\s+/);
-        svg.setAttribute('width', w);
-        svg.setAttribute('height', h);
-
-        const serializer = new XMLSerializer();
-        const newSource = serializer.serializeToString(doc);
-
-        return blobToImg(new Blob([newSource], { type: 'image/svg+xml' }));
-    }
-
+/**
+ * Get ImageData Object from blob
+ * @param {Blob} blob Blob to get data from
+ * @returns {Promise<ImageData>} Promise, which resolves into ImageData object
+ */
+export async function imageDataFromBlob (blob: Blob): Promise<ImageData> {
     // Prefer createImageBitmap as it's the off-thread option for Firefox.
     const drawable =
         'createImageBitmap' in self
             ? await createImageBitmap(blob)
             : await blobToImg(blob);
 
-    return drawable;
+    return imageDataFromDrawable(drawable);
 }
 
 /**
- * Get ImageData Object from file
- * @param {Blob} blob Blob to get data from
+ * Get ImageData Object from drawable
+ * @param {HTMLImageElement|ImageBitmap} drawable Blob to get data from
  * @returns {Promise<ImageData>} Promise, which resolves into ImageData object
  */
-export async function imageDataFromBlob (blob: Blob): Promise<ImageData> {
-    const drawable = await drawableFromBlob(blob);
-
+export async function imageDataFromDrawable (drawable: HTMLImageElement|ImageBitmap): Promise<ImageData> {
     const canvas = document.createElement('canvas');
     canvas.width = drawable.width;
     canvas.height = drawable.height;
